Reorganize into modules - Add note about public bot running latest release - Include release version - Unify command names with README
