+++ bot/cogs/__init__.py
@@ -1,0 +1,2 @@
+from .music_cog import Music
+from .server_assistant_cog import ServerAssistant
